+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Dropped support for Python 3.5-3.6 and will continue to support only current Python versions.
 
 ### Added:
-- optional attribution to map encoder/decoder output string back to input string (Issue #48)
+- optional attribution to map encoder/decoder output string back to input string (Issue #48, #79)
 
 ## v2.0.0 - 21.10.2021